LadderGame에 인터페이스를 활용한 의존성 주입을 통해 LadderCreator를 변경할 수 있게함
